requirements.txt
aiogram==3.12.0
asyncpg
apscheduler
python-dotenv
